+++ tests/createWebpackConfig-test.js
@@ -33,7 +33,7 @@ describe('createWebpackConfig()', () => {
   context('with only entry config', () => {
     let config = createWebpackConfig({entry: ['index.js']})
     it('creates a default webpack build config', () => {
-      expect(Object.keys(config)).toEqual(['module', 'output', 'plugins', 'resolve', 'resolveLoader', 'entry'])
+      expect(Object.keys(config)).toEqual(['module', 'output', 'plugins', 'resolve', 'entry'])
       expect(getLoaders(config.module.rules))
         .toContain('babel-loader')
         .toContain('extract-text-webpack-plugin')